app/requirements.txt
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
google-adk>=0.1.0 
